[update] move logger interface from usecase to interfaces
