+++ app/interfaces/category_controller.go
@@ -10,11 +10,11 @@ import (
 // A CategoryController is a controller for a comment.
 type CategoryController struct {
 	CategoryInteractor usecase.CategoryInteractor
-	Logger             usecase.Logger
+	Logger             Logger
 }
 
 // NewCategoryController creates a CategoryController.
-func NewCategoryController(connMySQL *sql.DB, logger usecase.Logger) *CategoryController {
+func NewCategoryController(connMySQL *sql.DB, logger Logger) *CategoryController {
 	return &CategoryController{
 		CategoryInteractor: usecase.CategoryInteractor{
 			CategoryRepository: &CategoryRepository{
